refactored controllers tests
